backend/requirements.txt
# Copyright (c) 2025 Bhagya Dissanayake
# All rights reserved. This code is proprietary and confidential.
# Unauthorized copying, distribution, or use is strictly prohibited.

fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
redis==5.0.1
prometheus-client==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
